pyproject.toml
[project]
name = "linux-remote-tool"
version = "1.0.1"
description = "Linux remote runtime and bridge for AI agent skills"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "paramiko>=3.5.0",
    "pydantic>=2.9.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.1",
]

[project.scripts]
lr = "linux_remote_tool.cli:main"
lr-codex = "linux_remote_tool.codex_bridge:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0"
]

[dependency-groups]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "--cov=src/linux_remote_tool --cov-report=term-missing"
python_files = "test_*.py"
testpaths = ["tests"]
